server/requirements.txt
fastapi
python-multipart
uvicorn
python-dotenv
google-genai
pydantic
transformers
chromadb
Pillow
matplotlib
scikit-learn
plotly
aiohttp
aiofiles
spacy
networkx
open_clip_torch
boto3
